+++ cedar-policy-core/src/evaluator.rs
@@ -406,7 +406,11 @@ impl<'q, 'e> Evaluator<'e> {
                 match op {
                     BinaryOp::Eq => Ok((arg1 == arg2).into()),
                     // comparison and arithmetic operators, which only work on Longs
-                    BinaryOp::Less | BinaryOp::LessEq | BinaryOp::Add | BinaryOp::Sub => {
+                    BinaryOp::Less
+                    | BinaryOp::LessEq
+                    | BinaryOp::Add
+                    | BinaryOp::Sub
+                    | BinaryOp::Mul => {
                         let i1 = arg1.get_as_long()?;
                         let i2 = arg2.get_as_long()?;
                         match op {
@@ -424,6 +428,16 @@ impl<'q, 'e> Evaluator<'e> {
                             },
                             BinaryOp::Sub => match i1.checked_sub(i2) {
                                 Some(diff) => Ok(diff.into()),
+                                None => Err(EvaluationError::IntegerOverflow(
+                                    IntegerOverflowError::BinaryOp {
+                                        op: *op,
+                                        arg1,
+                                        arg2,
+                                    },
+                                )),
+                            },
+                            BinaryOp::Mul => match i1.checked_mul(i2) {
+                                Some(prod) => Ok(prod.into()),
                                 None => Err(EvaluationError::IntegerOverflow(
                                     IntegerOverflowError::BinaryOp {
                                         op: *op,
@@ -515,21 +529,6 @@ impl<'q, 'e> Evaluator<'e> {
                     }
                 }
             }
-            ExprKind::MulByConst { arg, constant } => match self.partial_interpret(arg, slots)? {
-                PartialValue::Value(arg) => {
-                    let i1 = arg.get_as_long()?;
-                    match i1.checked_mul(*constant) {
-                        Some(prod) => Ok(prod.into()),
-                        None => Err(EvaluationError::IntegerOverflow(
-                            IntegerOverflowError::Multiplication {
-                                arg,
-                                constant: *constant,
-                            },
-                        )),
-                    }
-                }
-                PartialValue::Residual(r) => Ok(PartialValue::Residual(Expr::mul(r, *constant))),
-            },
             ExprKind::ExtensionFunctionApp { fn_name, args } => {
                 let args = args
                     .iter()
@@ -862,6 +861,8 @@ fn stack_size_check() -> Result<()> {
 #[cfg(test)]
 pub mod test {
     use std::str::FromStr;
+
+    use cool_asserts::assert_matches;
 
     use super::*;
 
@@ -2582,30 +2583,33 @@ pub mod test {
         );
         // 5 * (-3)
         assert_eq!(
-            eval.interpret_inline_policy(&Expr::mul(Expr::val(5), -3)),
-            Ok(Value::Lit(Literal::Long(-15)))
+            eval.interpret_inline_policy(&Expr::mul(Expr::val(5), Expr::val(-3))),
+            Ok(Value::from(-15))
         );
         // 5 * 0
         assert_eq!(
-            eval.interpret_inline_policy(&Expr::mul(Expr::val(5), 0)),
-            Ok(Value::Lit(Literal::Long(0)))
+            eval.interpret_inline_policy(&Expr::mul(Expr::val(5), Expr::val(0))),
+            Ok(Value::from(0))
         );
         // "5" * 0
-        assert_eq!(
-            eval.interpret_inline_policy(&Expr::mul(Expr::val("5"), 0)),
-            Err(EvaluationError::TypeError {
-                expected: vec![Type::Long],
-                actual: Type::String
-            })
+        assert_matches!(
+            eval.interpret_inline_policy(&Expr::mul(Expr::val("5"), Expr::val(0))),
+            Err(e) => assert_eq!(e,
+                EvaluationError::TypeError {
+                    expected: vec![Type::Long],
+                    actual: Type::String
+                }
+            )
         );
         // overflow
         assert_eq!(
-            eval.interpret_inline_policy(&Expr::mul(Expr::val(std::i64::MAX - 1), 3)),
+            eval.interpret_inline_policy(&Expr::mul(Expr::val(i64::MAX - 1), Expr::val(3))),
             Err(EvaluationError::IntegerOverflow(
-                IntegerOverflowError::Multiplication {
-                    arg: Value::from(std::i64::MAX - 1),
-                    constant: 3,
-                }
+                IntegerOverflowError::BinaryOp {
+                    op: BinaryOp::Mul,
+                    arg1: Value::from(i64::MAX - 1),
+                    arg2: Value::from(3),
+                },
             ))
         );
     }
@@ -4916,7 +4920,7 @@ pub mod test {
         let exts = Extensions::none();
         let eval = Evaluator::new(&empty_request(), &es, &exts).unwrap();
 
-        let e = Expr::mul(Expr::unknown("a"), 32);
+        let e = Expr::mul(Expr::unknown("a"), Expr::val(32));
         let r = eval.partial_interpret(&e, &HashMap::new()).unwrap();
         assert_eq!(r, PartialValue::Residual(e));
     }