+++ cedar-policy-core/src/evaluator/err.rs
@@ -127,7 +127,8 @@ fn pretty_type_error(expected: &[Type], actual: &Type) -> String {
 
 #[derive(Debug, PartialEq, Clone, Error)]
 pub enum IntegerOverflowError {
-    #[error("integer overflow while attempting to {} the values {arg1} and {arg2}", match .op { BinaryOp::Add => "add", BinaryOp::Sub => "subtract", _ => "perform an operation on" })]
+    /// Overflow during a binary operation
+    #[error("integer overflow while attempting to {} the values {arg1} and {arg2}", match .op { BinaryOp::Add => "add", BinaryOp::Sub => "subtract", BinaryOp::Mul => "multiply", _ => "perform an operation on" })]
     BinaryOp {
         /// overflow while evaluating this operator
         op: BinaryOp,
@@ -135,14 +136,6 @@ pub enum IntegerOverflowError {
         arg1: Value,
         /// second argument to that operator
         arg2: Value,
-    },
-
-    #[error("integer overflow while attempting to multiply {arg} by {constant}")]
-    Multiplication {
-        /// first argument, which wasn't necessarily a constant in the policy
-        arg: Value,
-        /// second argument, which was a constant in the policy
-        constant: i64,
     },
 
     /// Overflow during an integer negation operation