+++ cedar-policy/CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Validation error messages render types in the new, more readable, schema
   syntax. (#708, resolving #242)
 - Removed unnecessary lifetimes from some validation related structs (#715)
+- Return specific error type instead of using `impl miette::Diagnostic` (#724, resolving #723)
 
 ## [3.1.1] - 2024-03-14